programming ex9
I copied the code from ex 6 and added try and exception handlers. I have the IO and Value error handlers print the exception object related to the exception. To test these I opened a file that doesn't exist in read mode and I tried to convert a string to an int that has a word instead of a number.