programming ex12
I created a txt file that is 12 lines long and has the days on the months on it. 31,28, ect. I also created a file that has steps for 365 days in it. I then created a loop that runs on the number of days in a month from that file and then averages them and moves to the next line in the month file for the days in a month